+++ src/main/java/raven/datetime/component/date/PanelDate.java
@@ -65,7 +65,10 @@ public class PanelDate extends JPanel {
     }
 
     protected void createDateHeader() {
-        String weekdays[] = DateFormatSymbols.getInstance().getShortWeekdays();
+        String[] weekdays = DatePicker.getDefaultWeekdays();
+        if (weekdays == null) {
+            weekdays = DateFormatSymbols.getInstance().getShortWeekdays();
+        }
         // swap monday to the start day of week
         if (datePicker.isStartWeekOnMonday()) {
             String sunday = weekdays[1];