+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - DatePicker:
     - Add new option `defaultDateCellRenderer` to custom date cell renderer (issue #18) and (PR #19)
+    - Add global `DatePicker.defaultWeekdays` for change the default label weekdays (issue #20)
 
 ### Fixed bugs